refactor(pkg): extract shared servicediscovery package :hammer:
De-duplicate the identical Service Discovery bootstrap code from the ledger
and crm components into pkg/servicediscovery (BuildManager, ParseServerPort,
BuildServiceDescriptor, ResolveAuthHost, Runnable). Both components now import it.
Also applies two review fixes: gate ParseServerPort/descriptor behind the enabled
flag (strict no-op parity when SD_ENABLED=false) and create the resolve timeout
context only when auth is enabled.

X-Lerian-Ref: 0x1
