+++ components/crm/internal/bootstrap/service.go
@@ -6,13 +6,9 @@ package bootstrap
 
 import (
 	"context"
-	"fmt"
-	"net"
 	"os"
 	"os/signal"
-	"strconv"
 	"syscall"
-	"time"
 
 	libCommons "github.com/LerianStudio/lib-commons/v5/commons"
 	tmevent "github.com/LerianStudio/lib-commons/v5/commons/tenant-manager/event"
@@ -23,17 +19,8 @@ import (
 	"github.com/LerianStudio/midaz/v3/components/crm/internal/services/encryption"
 	"github.com/LerianStudio/midaz/v3/pkg/crypto"
 	"github.com/LerianStudio/midaz/v3/pkg/crypto/kms/vault"
+	pkgsd "github.com/LerianStudio/midaz/v3/pkg/servicediscovery"
 )
-
-// serviceDiscoveryDeregisterTimeout bounds the shutdown deregister call so a
-// slow or unreachable registry cannot hold the process open at exit. TTL expiry
-// is the backstop when deregister does not complete in time.
-const serviceDiscoveryDeregisterTimeout = 5 * time.Second
-
-// serviceDiscoveryResolveTimeout bounds the plugin-auth host resolution at boot
-// so a slow or unreachable registry cannot stall startup. On timeout the caller
-// degrades to the static auth address.
-const serviceDiscoveryResolveTimeout = 5 * time.Second
 
 // Service is the application glue where we put all top level components to be used.
 type Service struct {
@@ -66,8 +53,8 @@ type Service struct {
 	// ServiceDiscovery is the SD Manager (always non-nil; a working no-op when
 	// discovery is disabled). ServiceDiscoveryEnabled mirrors SD_ENABLED so Run()
 	// can decide whether to wire a register/deregister runnable. ServiceDescriptor
-	// is the descriptor this service advertises; it is populated in a later task
-	// and left zero-value here.
+	// is the descriptor this service advertises; it is built at wiring time only
+	// when discovery is enabled and left zero-value otherwise.
 	ServiceDiscovery        *libsd.Manager
 	ServiceDiscoveryEnabled bool
 	ServiceDescriptor       libsd.Service
@@ -108,11 +95,10 @@ func (app *Service) launcherApps() []launcherApp {
 	// best-effort only; the 30s TTL on the descriptor is the real guarantee that
 	// a stale instance drops out of the registry if deregister is missed.
 	if app.ServiceDiscoveryEnabled {
-		apps = append(apps, launcherApp{"Service Discovery", &serviceDiscoveryRunnable{
-			manager: app.ServiceDiscovery,
-			svc:     app.ServiceDescriptor,
-			logger:  app.Logger,
-		}})
+		apps = append(apps, launcherApp{
+			"Service Discovery",
+			pkgsd.NewRunnable(app.ServiceDiscovery, app.ServiceDescriptor, app.Logger),
+		})
 	}
 
 	apps = append(apps, launcherApp{"HTTP Service", app.Server})
@@ -135,86 +121,6 @@ func (app *Service) launcherApps() []launcherApp {
 	}
 
 	return apps
-}
-
-// parseServerPort extracts the numeric port from a listen address. It accepts
-// both the leading-colon form (":4003") and the host:port form
-// ("0.0.0.0:8080"); net.SplitHostPort handles both. A malformed address is a
-// config bug and surfaces as an error for fail-fast handling at wiring time.
-func parseServerPort(serverAddress string) (int, error) {
-	_, portStr, err := net.SplitHostPort(serverAddress)
-	if err != nil {
-		return 0, fmt.Errorf("parsing server address %q: %w", serverAddress, err)
-	}
-
-	return strconv.Atoi(portStr)
-}
-
-// buildCRMServiceDescriptor builds the registry descriptor for this CRM
-// instance. Address and Scheme are intentionally left unset: Manager.Register
-// fills them from SD_ADVERTISE_ADDRESS. The TTL health check needs no reachable
-// HTTP endpoint — the registry heartbeats from inside the process.
-func buildCRMServiceDescriptor(port int) libsd.Service {
-	return libsd.Service{
-		ID:          "midaz-crm-" + strconv.Itoa(port),
-		Name:        "midaz-crm",
-		Port:        port,
-		HealthCheck: &libsd.HealthCheck{TTL: "30s"},
-	}
-}
-
-// serviceDiscoveryRunnable adapts service-discovery register/deregister to the
-// libCommons.App interface. It registers asynchronously at start, blocks until
-// SIGINT/SIGTERM, then deregisters on shutdown. A deregister failure is logged
-// at Warn but not propagated: TTL expiry is the backstop and the Launcher cannot
-// meaningfully react at shutdown.
-type serviceDiscoveryRunnable struct {
-	manager *libsd.Manager
-	svc     libsd.Service
-	logger  libLog.Logger
-
-	// notifyContext builds the signal-scoped context that gates the runnable's
-	// lifetime. It defaults to signal.NotifyContext in production; tests inject a
-	// factory returning a context they cancel to simulate SIGTERM deterministically.
-	notifyContext func(context.Context, ...os.Signal) (context.Context, context.CancelFunc)
-}
-
-// Run registers the service asynchronously against the signal-scoped context,
-// blocks until SIGINT/SIGTERM, then deregisters under a fresh short-lived
-// context (the signal context is already cancelled by then).
-func (r *serviceDiscoveryRunnable) Run(_ *libCommons.Launcher) error {
-	if r == nil || r.manager == nil {
-		return nil
-	}
-
-	notify := r.notifyContext
-	if notify == nil {
-		notify = signal.NotifyContext
-	}
-
-	sigCtx, stop := notify(context.Background(), os.Interrupt, syscall.SIGTERM)
-	defer stop()
-
-	// RegisterAsync is non-blocking and retries in the background until sigCtx
-	// is cancelled. It needs the app-lifetime (signal) context, not a
-	// request-scoped one.
-	r.manager.RegisterAsync(sigCtx, r.svc)
-
-	<-sigCtx.Done()
-
-	ctx, cancel := context.WithTimeout(context.Background(), serviceDiscoveryDeregisterTimeout)
-	defer cancel()
-
-	if err := r.manager.Deregister(ctx, r.svc.ID); err != nil && r.logger != nil {
-		r.logger.Log(
-			context.Background(), libLog.LevelWarn,
-			"service discovery deregister returned error",
-			libLog.String("service_id", r.svc.ID),
-			libLog.Err(err),
-		)
-	}
-
-	return nil
 }
 
 // streamingProducerRunnable adapts the lib-streaming Producer's Close hook to
