+++ components/ledger/internal/bootstrap/service.go
@@ -6,13 +6,9 @@ package bootstrap
 
 import (
 	"context"
-	"fmt"
-	"net"
 	"os"
 	"os/signal"
-	"strconv"
 	"syscall"
-	"time"
 
 	libCommons "github.com/LerianStudio/lib-commons/v5/commons"
 	tmconsumer "github.com/LerianStudio/lib-commons/v5/commons/tenant-manager/consumer"
@@ -21,18 +17,8 @@ import (
 	"github.com/LerianStudio/lib-observability/metrics"
 	libOpentelemetry "github.com/LerianStudio/lib-observability/tracing"
 	libsd "github.com/LerianStudio/lib-service-discovery"
+	pkgsd "github.com/LerianStudio/midaz/v3/pkg/servicediscovery"
 )
-
-// serviceDiscoveryDeregisterTimeout bounds the shutdown deregister call so a
-// slow or unreachable registry cannot hold the process open at exit. TTL expiry
-// is the backstop when deregister does not complete in time.
-const serviceDiscoveryDeregisterTimeout = 5 * time.Second
-
-// serviceDiscoveryResolveTimeout bounds the boot-time plugin-auth resolve so a
-// TCP-reachable but slow/hung registry (brownout) cannot stall boot. On deadline
-// the resolve degrades to the static auth host, keeping "a discovery outage never
-// stalls boot" true for the slow-registry case, not just connection refused.
-const serviceDiscoveryResolveTimeout = 5 * time.Second
 
 // Service is the unified ledger service that owns all infrastructure directly.
 type Service struct {
@@ -67,8 +53,9 @@ type Service struct {
 	// to register the discovery register/deregister Launcher app.
 	ServiceDiscoveryEnabled bool
 	// ServiceDescriptor is the descriptor advertised to the registry. It is
-	// built once at wiring time (so a malformed SERVER_ADDRESS fails fast
-	// there) and reused by the service-discovery runnable.
+	// built at wiring time only when discovery is enabled (so a malformed
+	// SERVER_ADDRESS never aborts boot with discovery off) and reused by the
+	// service-discovery runnable. It is zero-value when discovery is disabled.
 	ServiceDescriptor libsd.Service
 }
 
@@ -107,11 +94,10 @@ func (s *Service) launcherApps() []launcherApp {
 	// best-effort only; the 30s TTL on the descriptor is the real guarantee that
 	// a stale instance drops out of the registry if deregister is missed.
 	if s.ServiceDiscoveryEnabled {
-		apps = append(apps, launcherApp{"Service Discovery", &serviceDiscoveryRunnable{
-			manager: s.ServiceDiscovery,
-			svc:     s.ServiceDescriptor,
-			logger:  s.Logger,
-		}})
+		apps = append(apps, launcherApp{
+			"Service Discovery",
+			pkgsd.NewRunnable(s.ServiceDiscovery, s.ServiceDescriptor, s.Logger),
+		})
 	}
 
 	apps = append(apps, launcherApp{"Unified HTTP Server", s.UnifiedServer})
@@ -200,86 +186,6 @@ func (r *streamingProducerRunnable) Run(_ *libCommons.Launcher) error {
 			context.Background(), libLog.LevelWarn,
 			"streaming producer Close returned error",
 			libLog.String("error", err.Error()),
-		)
-	}
-
-	return nil
-}
-
-// parseServerPort extracts the numeric port from a listen address. It accepts
-// both the leading-colon form (":3002") and the host:port form
-// ("0.0.0.0:8080"); net.SplitHostPort handles both. A malformed address is a
-// config bug and surfaces as an error for fail-fast handling at wiring time.
-func parseServerPort(serverAddress string) (int, error) {
-	_, portStr, err := net.SplitHostPort(serverAddress)
-	if err != nil {
-		return 0, fmt.Errorf("parsing server address %q: %w", serverAddress, err)
-	}
-
-	return strconv.Atoi(portStr)
-}
-
-// buildLedgerServiceDescriptor builds the registry descriptor for this ledger
-// instance. Address and Scheme are intentionally left unset: Manager.Register
-// fills them from SD_ADVERTISE_ADDRESS. The TTL health check needs no reachable
-// HTTP endpoint — the registry heartbeats from inside the process.
-func buildLedgerServiceDescriptor(port int) libsd.Service {
-	return libsd.Service{
-		ID:          "midaz-ledger-" + strconv.Itoa(port),
-		Name:        "midaz-ledger",
-		Port:        port,
-		HealthCheck: &libsd.HealthCheck{TTL: "30s"},
-	}
-}
-
-// serviceDiscoveryRunnable adapts service-discovery register/deregister to the
-// libCommons.App interface. It registers asynchronously at start, blocks until
-// SIGINT/SIGTERM, then deregisters on shutdown. A deregister failure is logged
-// at Warn but not propagated: TTL expiry is the backstop and the Launcher cannot
-// meaningfully react at shutdown.
-type serviceDiscoveryRunnable struct {
-	manager *libsd.Manager
-	svc     libsd.Service
-	logger  libLog.Logger
-
-	// notifyContext builds the signal-scoped context that gates the runnable's
-	// lifetime. It defaults to signal.NotifyContext in production; tests inject a
-	// factory returning a context they cancel to simulate SIGTERM deterministically.
-	notifyContext func(context.Context, ...os.Signal) (context.Context, context.CancelFunc)
-}
-
-// Run registers the service asynchronously against the signal-scoped context,
-// blocks until SIGINT/SIGTERM, then deregisters under a fresh short-lived
-// context (the signal context is already cancelled by then).
-func (r *serviceDiscoveryRunnable) Run(_ *libCommons.Launcher) error {
-	if r == nil || r.manager == nil {
-		return nil
-	}
-
-	notify := r.notifyContext
-	if notify == nil {
-		notify = signal.NotifyContext
-	}
-
-	sigCtx, stop := notify(context.Background(), os.Interrupt, syscall.SIGTERM)
-	defer stop()
-
-	// RegisterAsync is non-blocking and retries in the background until sigCtx
-	// is cancelled. It needs the app-lifetime (signal) context, not a
-	// request-scoped one.
-	r.manager.RegisterAsync(sigCtx, r.svc)
-
-	<-sigCtx.Done()
-
-	ctx, cancel := context.WithTimeout(context.Background(), serviceDiscoveryDeregisterTimeout)
-	defer cancel()
-
-	if err := r.manager.Deregister(ctx, r.svc.ID); err != nil && r.logger != nil {
-		r.logger.Log(
-			context.Background(), libLog.LevelWarn,
-			"service discovery deregister returned error",
-			libLog.String("service_id", r.svc.ID),
-			libLog.Err(err),
 		)
 	}
